backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.5